Attempting to convert project to using Stack.
I ran the following commands:
    stack init --resolver lts-5.1
    stack setup
    stack build

Initialising stack fails as some of the dependencies are old and don't exist in the stackage version chosen. I updated these in the light-haskell.cabal file. The error message I was getting is below:

Selected resolver 'lts-5.1' does not have all the packages to match your requirements.
    aeson version 0.9.0.1 found
        - light-haskell requires ==0.7.*
    haskell-src-exts version 1.17.1 found
        - light-haskell requires ==1.14.*
    text version 1.2.2.0 found
        - light-haskell requires ==1.1.*

Then, when building there is a typing error for getLintIssues() in LTHaskellClient. For the time being, I have set this method to undefined to allow the build to pass.